port raid1 hashed barrier implementation to raid10
This is a port of Coly's md/raid1 barrier re-implementation to raid10.  It's
based on work done by Lidong Zhong
(https://www.spinics.net/lists/raid/msg60675.html), with a bugfix and some
restrictions.

In addition to Lidong's work, this commit includes a sysfs-adjustable resync
depth: as it was, raid10 resync could generate an I/O depth as large as 512,
behind which user I/O waited.  Perhaps this was necessary on spinning media,
but on flash devices, there seems little reason to jam the queue so full.  New
default depth is 32, matching raid1.

raise_barrier now waits for up to 8 pending I/O's to complete before
incrementing the barrier.  This shifts some burden onto resync/recovery to
avoid jamming a barrer in the middle of a user workload.

Bugfix to Lidong's work:

- The handling of raise_barrier "force" in raid10 recovery could end up forcing
  a barrier in situations where it shouldn't have, when multiple disks were not
  in sync (e.g., in a 5+ disk array).  I introduced a function
  (raid10_should_force_barrier) that walks the r10bio list to determine if the
  current loop had barriered the region already.

Limitations:

- I was not confident in the split between read and write barriers for raid10.
  The code here has taken the conservative approach of a single barrier type,
  used for both.

- Reshape is not supported.  We don't require it, and the way it treated the
  barrier had complex interactions.

- If the raid10 chunk size exceeds the BARRIER_UNIT (64MiB), the code doesn't
  correctly barrier two regions.  I simply rejected chunk sizes larger than the
  barrier unit.