raid10: use logical bio for raising the barrier in handle_read_error
This fixes a bug that was introduced with the hashed barrier support for
raid10.  If md handled a read error, it would use the physical device address
to raise the read's barrier, instead of the logical address.  The bug would
leave a barrier raised for the logical address, and a negative refcount for the
barrier at the physical address.  Later, a resync or check would hang the
kernel thread, waiting indefinitely for the barrier.  This fix correctly uses
the logical address to raise the barrier when handling a read error.